+++ c/src/gigperformer/sdk/imports.c
@@ -37,6 +37,8 @@ extern "C"
     DeclareFunc(GP_GetPluginParameterCount);
     DeclareFunc(GP_GetPluginParameterName);
     DeclareFunc(GP_GetPluginParameterText);
+    DeclareFunc(GP_GetPluginCaption);
+    DeclareFunc(GP_GetPluginName);
 
     DeclareFunc(GP_SetWidgetHideOnPresentation);
     DeclareFunc(GP_GetWidgetHideState);
@@ -196,6 +198,8 @@ extern "C"
         R(GP_GetPluginParameterCount);
         R(GP_GetPluginParameterName);
         R(GP_GetPluginParameterText);
+        R(GP_GetPluginCaption);
+        R(GP_GetPluginName);
 
         R(GP_MapWidgetToPluginParameter);
         R(GP_GetWidgetList);